Additional improvements on the generalized add/edit app forms.
1897f0aa8
