+++ src/Entity/Form/DeveloperAppCreateFormForDeveloper.php
@@ -19,73 +19,18 @@
 
 namespace Drupal\apigee_edge\Entity\Form;
 
-use Drupal\apigee_edge\Entity\ApiProductInterface;
-use Drupal\apigee_edge\Entity\Controller\AppCredentialControllerInterface;
-use Drupal\apigee_edge\Entity\Controller\DeveloperAppCredentialControllerFactoryInterface;
 use Drupal\apigee_edge\Entity\DeveloperStatusCheckTrait;
-use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\Url;
-use Drupal\user\UserInterface;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * Dedicated form handler that allows a developer to create an developer app.
  */
-class DeveloperAppCreateFormForDeveloper extends AppForm {
+class DeveloperAppCreateFormForDeveloper extends DeveloperAppFormForDeveloper {
 
   use DeveloperStatusCheckTrait;
   use AppCreateFormTrait;
   use DeveloperAppFormTrait;
-
-  /**
-   * The user from the route.
-   *
-   * Use getUser() instead.
-   *
-   * @var \Drupal\user\UserInterface
-   */
-  protected $user;
-
-  /**
-   * The app credential controller factory.
-   *
-   * @var \Drupal\apigee_edge\Entity\Controller\DeveloperAppCredentialControllerFactoryInterface
-   */
-  protected $appCredentialControllerFactory;
-
-  /**
-   * DeveloperAppCreateFormForDeveloper constructor.
-   *
-   * @param \Drupal\apigee_edge\Entity\Controller\DeveloperAppCredentialControllerFactoryInterface $app_credential_controller_factory
-   *   The app credential controller factory.
-   * @param \Drupal\Core\Entity\EntityTypeManagerInterface $entity_type_manager
-   *   Entity type manager.   *.
-   */
-  public function __construct(DeveloperAppCredentialControllerFactoryInterface $app_credential_controller_factory, EntityTypeManagerInterface $entity_type_manager) {
-    parent::__construct($entity_type_manager);
-    $this->appCredentialControllerFactory = $app_credential_controller_factory;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function create(ContainerInterface $container) {
-    return new static(
-      $container->get('apigee_edge.controller.developer_app_credential_factory'),
-      $container->get('entity_type.manager')
-    );
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function buildForm(array $form, FormStateInterface $form_state, UserInterface $user = NULL) {
-    // This is the only place where we can grab additional route parameters.
-    // See implementation in parent.
-    $this->user = $user;
-    return parent::buildForm($form, $form_state);
-  }
 
   /**
    * {@inheritdoc}
@@ -109,35 +54,6 @@ class DeveloperAppCreateFormForDeveloper extends AppForm {
   protected function getRedirectUrl(): Url {
     $entity = $this->getEntity();
     return $entity->toUrl('collection-by-developer');
-  }
-
-  /**
-   * The user from the route or the current user if it is not available.
-   *
-   * @return \Drupal\user\UserInterface
-   *   User object.
-   */
-  protected function getUser(): UserInterface {
-    return $this->user ?? $this->entityTypeManager->getStorage('user')->load(\Drupal::currentUser()->id());
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  protected function apiProductList(): array {
-    $api_products = parent::apiProductList();
-    array_filter($api_products, function (ApiProductInterface $product) {
-      return $product->access('assign', $this->getUser());
-    });
-
-    return $api_products;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  protected function appCredentialController(string $owner, string $app_name): AppCredentialControllerInterface {
-    return $this->appCredentialControllerFactory->developerAppCredentialController($owner, $app_name);
   }
 
 }
