+++ src/Entity/Form/DeveloperAppEditFormForDeveloper.php
@@ -20,13 +20,32 @@
 
 namespace Drupal\apigee_edge\Entity\Form;
 
+use Drupal\apigee_edge\Entity\Controller\AppCredentialControllerInterface;
+use Drupal\apigee_edge\Entity\DeveloperStatusCheckTrait;
 use Drupal\Core\Form\FormStateInterface;
+use Drupal\Core\Routing\RouteMatchInterface;
 use Drupal\Core\Url;
 
 /**
  * Dedicated form handler that allows a developer to edit its developer app.
  */
-class DeveloperAppEditFormForDeveloper extends DeveloperAppEditForm {
+class DeveloperAppEditFormForDeveloper extends DeveloperAppFormForDeveloper {
+
+  use DeveloperStatusCheckTrait;
+  use AppEditFormTrait;
+  use DeveloperAppFormTrait;
+
+  /**
+   * {@inheritdoc}
+   */
+  public function form(array $form, FormStateInterface $form_state) {
+    $form = parent::form($form, $form_state);
+    /** @var \Drupal\apigee_edge\Entity\DeveloperAppInterface $app */
+    $app = $this->entity;
+    $this->checkDeveloperStatus($app->getOwnerId());
+
+    return $form;
+  }
 
   /**
    * {@inheritdoc}
@@ -44,6 +63,26 @@ class DeveloperAppEditFormForDeveloper extends DeveloperAppEditForm {
     $actions['delete']['#url'] = $this->getEntity()->toUrl('delete-form-for-developer');
 
     return $actions;
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getEntityFromRouteMatch(RouteMatchInterface $route_match, $entity_type_id) {
+    if ($route_match->getRawParameter('app') !== NULL) {
+      $entity = $route_match->getParameter('app');
+    }
+    else {
+      $entity = parent::getEntityFromRouteMatch($route_match, $entity_type_id);
+    }
+    return $entity;
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  protected function appCredentialController(string $owner, string $app_name): AppCredentialControllerInterface {
+    return $this->appCredentialControllerFactory->developerAppCredentialController($owner, $app_name);
   }
 
 }
